pyproject.toml
[tool.poetry]
name = "aleksbel_gen_diff"
version = "0.1.3"
description = ""
authors = ["Aleks Bel <aleksb2001@mail.ru>"]
packages = [
    {include = "gen_diff" },
  ]

[tool.poetry.dependencies]
python = "^3.6"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
pytest = "^5.3"
pytest-cov = "^2.8"
pyyaml = "^5.3"

[tool.poetry.scripts]
gendiff = "gen_diff.scripts.gendiff:main"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

